Backend/requirements.txt
flask>=3.0.0
flask-pymongo>=2.3.0
flask-cors>=4.0.0
flask-bcrypt>=1.0.1
flask-jwt-extended>=4.6.0
pymongo>=4.6.0
python-dotenv>=1.0.0
openai>=1.0.0
